back_end/requirement.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
flask-mail==0.9.1
werkzeug==2.3.7
itsdangerous==2.1.2
mysqlclient==2.2.1
PyMySQL==1.1.0
opencv-python==4.9.0.80
numpy==1.26.4
imutils==0.5.4
PyMuPDF==1.24.0
pytesseract==0.3.13
Pillow==10.3.0
pandas==2.2.0
reportlab==4.1.0
PyJWT==2.8.0
